# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sidekeek API Service",
    "description": "Group of services provided to maintain sidekeek-client needs",
    "contact": "yourname@something.com",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/sideAPIkeek",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/editProductDesc": {
      "post": {
        "tags": ["ProductController"],
        "summary": "edit product description",
        "description": "edit specific product description",
        "operationId": "editProductDesc",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "idProduct",
            "in" : "formData",
            "description": "idProduct",
            "required": true,
            "type": "string"
          },
          {
            "name": "namaProduk",
            "in" : "formData",
            "description": "namaProduk",
            "required": false,
            "type": "string"
          },
          {
            "name": "harga",
            "in" : "formData",
            "description": "harga",
            "required": false,
            "type": "string"
          },
          {
            "name": "productDesc",
            "in" : "formData",
            "description": "productDesc",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/editProductDesc"
            }
          }
        }
      }
    },
    "/editProductImage": {
      "post": {
        "tags": ["ProductController"],
        "summary": "edit product image",
        "description": "edit specific product-image image",
        "operationId": "editProductImage",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "idProductImage",
            "in" : "formData",
            "description": "idProductImage",
            "required": true,
            "type": "string"
          },
          {
            "name": "imgbase64",
            "in" : "formData",
            "description": "imgbase64",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/editProductImage"
            }
          }
        }
      }
    },
...
    "/addHostReview": {
      "post": {
        "tags": ["ProfileController"],
        "summary": "add Host Review",
        "description": "add Host Review",
        "operationId": "addHostReview",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "score",
            "in" : "formData",
            "description": "score",
            "required": true,
            "type": "string"
          },
          {
            "name": "idHost",
            "in" : "formData",
            "description": "idHost",
            "required": true,
            "type": "string"
          },
          {
            "name": "comment",
            "in" : "formData",
            "description": "comment",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in" : "formData",
            "description": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/post-simple"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "post-simple": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
    "editProductImage": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
    "editProductDesc": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
  }
}